+++ src/helpers/parse-keyword-value.ts
@@ -11,6 +11,8 @@
  */
 export const parseKeywordValue = (rawValue: string): string => {
     // !IMPORTANT: Do not move constants outside of this function
+    // Remember to add new keywords to "isCookieSetWithValue" function in "cookie-utils", if required
+    // https://github.com/AdguardTeam/Scriptlets/issues/489
     const NOW_VALUE_KEYWORD = '$now$';
     const CURRENT_DATE_KEYWORD = '$currentDate$';
     const CURRENT_ISO_DATE_KEYWORD = '$currentISODate$';